Refactor selectors into several files
