+++ app/javascript/app/pages/national-context/historical-emissions/historical-emissions-selectors/historical-emissions-selectors.js
@@ -1,0 +1,22 @@
+import { createStructuredSelector } from 'reselect';
+import { getQuery } from './historical-emissions-get-selectors';
+import {
+  getSelectedOptions,
+  getFilterOptions,
+  getTop10EmittersOption
+} from './historical-emissions-filter-selectors';
+import { getEmissionParams } from './historical-emissions-fetch-selectors';
+import {
+  getModelSelected,
+  getChartData
+} from './historical-emissions-data-selectors';
+
+export const getGHGEmissions = createStructuredSelector({
+  selectedOptions: getSelectedOptions,
+  fieldToBreakBy: getModelSelected,
+  filterOptions: getFilterOptions,
+  query: getQuery,
+  emissionParams: getEmissionParams,
+  chartData: getChartData,
+  top10EmmitersOption: getTop10EmittersOption
+});
